+++ GithubBackup.Cli/Commands/Github/Auth/Pipeline/ILoginPipelineBuilder.cs
@@ -1,8 +1,0 @@
-﻿namespace GithubBackup.Cli.Commands.Github.Auth.Pipeline;
-
-internal interface ILoginPipelineBuilder
-{
-    public ILoginPipeline PersistedOnly();
-    public ILoginPipeline WithPersistent();
-    public ILoginPipeline WithoutPersistent();
-}