+++ GithubBackup.Cli/Commands/Github/Auth/Pipeline/ITokenFromConfigurationPipeline.cs
@@ -1,3 +1,0 @@
-﻿namespace GithubBackup.Cli.Commands.Github.Auth.Pipeline;
-
-internal interface ITokenFromConfigurationPipeline : ILoginPipeline { }