+++ GithubBackup.Cli/Commands/Github/Login/LoginArgDescriptions.cs
@@ -6,8 +6,8 @@ internal static class LoginArgDescriptions
         "Token",
         "Token",
         """
-        If not provided, the token will be aquired from the environment variable GITHUB_BACKUP_TOKEN.
-        If provided, device flow authentication will be ignored. Recommended for use on servers.");
+        If not provided, the token will be acquired from the environment variable GITHUB_BACKUP_TOKEN.
+        If provided, temporary cache and device flow authentication will be ignored. Recommended for use on servers.
         """
     );
 
@@ -16,7 +16,7 @@ internal static class LoginArgDescriptions
         "Device Flow Auth",
         """
         Interactive authentication using the device flow.
-        Requires a browser. Recommended for use on clients.
+        Bypasses the temporary token cache. Requires a browser. Recommended for use on clients.
         """
     );
 }