+++ GithubBackup.Cli/Commands/Github/Repositories/RepositoriesRunner.cs
@@ -30,12 +30,7 @@ internal sealed class RepositoriesRunner : ICommandRunner
 
     public async Task RunAsync(CancellationToken ct)
     {
-        await _loginService.WithPersistentAsync(
-            _globalArgs,
-            _repositoriesArgs.LoginArgs,
-            false,
-            ct
-        );
+        await _loginService.LoginAsync(_globalArgs, _repositoriesArgs.LoginArgs, ct);
 
         var options = new RepositoryOptions(
             _repositoriesArgs.Type,